cmd/juju/storage: add "list", tweak "show"
This commit introduces "juju storage list" to list
all of the storage instances in the environment.
The default output is a tabular format, similar to
"juju status --format tabular".

The existing "juju storage show" has been tweaked
to output as a map, rather than a slice. This makes
the output more like "juju status". We also translate
tags to IDs (unit/service name, storage ID), as tags
are not intended for users' eyes.

The StorageName value is now computed from the storage
ID. Finally, the "location", "available-size" and
"total-size" fields are omitted if unavailable. To
effect this, we make their types pointers, so they
are nullable.

This commit requires https://github.com/juju/names/pull/38